+++ R/initialize.R
@@ -1286,7 +1286,9 @@ load_tidy_iea_df <- function(.iea_file = sample_iea_data_path(),
   }
   if (apply_fixes) {
     out <- out |> 
-      fix_GHA_industry_electricity() |> 
+      # We no longer need to fix GHA's industry electricity, 
+      # because details are now available in the IEA's WEEB.
+      # fix_GHA_industry_electricity() |> 
       fix_GHA_psb() |> 
       fix_COL_electricity_generation()
   }